# backend/app/schemas/codegen.py
# generated by datamodel-codegen:
#   filename:  openapi.yml
#   timestamp: 2026-07-14T04:31:49+00:00

from __future__ import annotations

from typing import Any
from uuid import UUID

from pydantic import AwareDatetime, EmailStr, Field, constr

from app.schemas.base import BaseModelWithOrm


class RegisterResponse(BaseModelWithOrm):
    status: str = Field(..., title="Status")
    message: str = Field(..., title="Message")


class LogoutResponse(BaseModelWithOrm):
    message: str = Field(..., title="Message")


class LoginRequest(BaseModelWithOrm):
    identifier: str = Field(..., description="Email or Username", title="Identifier")
    password: constr(min_length=8) = Field(..., title="Password")


class ProfileResponse(BaseModelWithOrm):
    username: str = Field(..., title="Username")
    display_name: str | None = Field(None, title="Display Name")
    avatar_key: str | None = Field(None, title="Avatar Key")
    theme_preference: str | None = Field("default", title="Theme Preference")
    id: UUID = Field(..., title="Id")
    role: str = Field(..., title="Role")
    created_at: AwareDatetime = Field(..., title="Created At")


class ProjectResponse(BaseModelWithOrm):
    title: str = Field(..., title="Title")
    status: str | None = Field("pending", title="Status")
    duration_seconds: int | None = Field(None, title="Duration Seconds")
    started_at: AwareDatetime | None = Field(None, title="Started At")
    id: UUID = Field(..., title="Id")
    user_id: UUID = Field(..., title="User Id")
    created_at: AwareDatetime = Field(..., title="Created At")
    updated_at: AwareDatetime = Field(..., title="Updated At")
    folder_id: Any | None = Field(None, title="Folder Id")
    pinned_at: Any | None = Field(None, title="Pinned At")


class RefreshTokenRequest(BaseModelWithOrm):
    refresh_token: str = Field(..., title="Refresh Token")


class RegisterRequest(BaseModelWithOrm):
    email: EmailStr = Field(..., title="Email")
    username: constr(pattern=r"^[a-z0-9_-]+$", min_length=3, max_length=30) = Field(
        ..., title="Username"
    )
    password: constr(min_length=8) = Field(..., title="Password")


class Token(BaseModelWithOrm):
    access_token: str = Field(..., title="Access Token")
    refresh_token: str = Field(..., title="Refresh Token")
    token_type: str | None = Field("bearer", title="Token Type")
    user_id: str = Field(..., title="User Id")
    email: EmailStr = Field(..., title="Email")
    username: str = Field(..., title="Username")
    role: str = Field(..., title="Role")


class TokenRefreshResponse(BaseModelWithOrm):
    access_token: str = Field(..., title="Access Token")
    refresh_token: str = Field(..., title="Refresh Token")
    token_type: str | None = Field("bearer", title="Token Type")


class UserResponse(BaseModelWithOrm):
    id: UUID = Field(..., title="Id")
    email: EmailStr = Field(..., title="Email")
    profile: ProfileResponse | None = None
    created_at: AwareDatetime = Field(..., title="Created At")
    disabled_at: AwareDatetime | None = Field(None, title="Disabled At")


class ValidationError(BaseModelWithOrm):
    loc: list[str | int] = Field(..., title="Location")
    msg: str = Field(..., title="Message")
    type: str = Field(..., title="Error Type")
    input: Any | None = Field(None, title="Input")
    ctx: dict[str, Any] | None = Field(None, title="Context")


class HTTPValidationError(BaseModelWithOrm):
    detail: list[ValidationError] | None = Field(None, title="Detail")